[GitHub Usage] Moved ToKiloFormat to NumberExtensions class
